Downgrade flanagan.jar to 1.0.3 because 1.1.0 gives a strange error
[INFO] Compilation failure
could not parse error message: flanagan/math/FourierTransform.java(flanagan/math:FourierTransform.java):18: unmappable character for encoding UTF-8
*   UPDATES: 18 February 2006  method correlation correction (thanks to Daniel Mader, Universt�t Freiburg -- IMTEK)

I get the same error on two different machines, so I'm assuming
that this not a local misconfiguration. We're going to get rid of
this module anyway, so let's use the simplest solution for now.