Remove link messages_pl.xml -> messages_en.xml
Links are not supported on windows, and the content of the symbolic link
(the filename) was unsuccessfully parsed as XML.

In commit 2af227a882e90f167db92a6390911911ebac5f88 I intended to remove
this file according to the commit message, but created this link instead.
The "polish" messages are still equal to the english ones, so let's finally
remove this file.