Fixed license issues for TabbedPane component with cross.
Added the Oracle copyright notice to the class javadoc.
There is no such ready-to-use component in the standard
Java libraries, but there was an example of a TabbedPane
component with a cross for closing tabs in the Oracle
tutorials - it was better to use a ready solution than
to unnecessarily spend time to invent one from scratch.