Old monitor tags are automatically removed from the tag set.
Previously - if somebody would make an experiment that would last
for 24h, in the end of the experiment all the tags that were
received by Svarog throught the day would be stored in the
StyledMonitorTagSet and each drawing of tags would have to
browse through all of them to find those few that should be
rendered in the SignalView. Now - periodically a method is
run that deletes all tags that are older that the current
page size (+tag length).