Store generated-sources in a separate branch
It is cumbersome for macintosh and windows users to install protoc
to generate Java sources. They don't change often, but storing them
directly in one of the branches would add unnecessary noise.
Generated sources are stored in a separate orphan branch
(generated-sources) and can be checked out with
  util/generated-sources

After updating multiplexer.rules or protoc version, use
  util/generated-sources commit
to store a new version.

This patch doesn't actually do anything to pom.xml to tell it not to try to
regenerate generated sources.