Enabled preset managers to read one configuration from multiple files.
If one preset manager needs to hold presets which are stored in
multiple files, it is now possible by overriding
the AbstractPresetManager#getStandardFilename() method
to return the directory name from which all the files will
be read and included in the preset manager. This functionality
will be needed while reading EEG systems configurations.