EEG systems are now stored in the profile directory in separate file each.
EEG systems are stored in the eegSystems directory inside the profile
directory. If this directory does not exist (or contains no files)
it is created and predefined EEG systems from resource
org/signalml/app/config/eegSystems are copied into it.
This allows to change or add an EEG system without
the need of recompiling Svarog.