Removed the SeriousWarningDialog and all its references.
This dialog displayed a warning message and allowed the user
to press the Cancel or OK button after a specified timeout.
It was used when a user wanted to load a new preset,
save a preset, generate a montage etc. In places
where such a warning dialog is needed
(e.g. overwriting an existsing preset), it was replaced
by the ErrorsDialog#showWarningYesNoDialog which is
simpler to use and less frustrating.