Fix-error while copying default EEG systems using a Svarog JAR snapshot.
There was a bug which caused that EEG systems resource files were not
found when running the svarog-0.5.0-SNAPSHOT-onejar.jar (trac #163).
Listing all files in a resource-directory worked for 'mvn exec:java'
but did not work when running a packaged JAR. Implementing a resource
search that worked for both was problematic, so now the list of files
that should be copied is stored inside of the EegSystemsPresetManager.