Added more EEG systems: EEG 10_20 (Cap19, 25), EEG 10_10 (Cap 33 and 47).
Also: the way the 3D polar positions of electrodes were mapped on a 2D
plane had to be changed. Previously the electrodes were rendered as if
the head is seen from above. This resulted in electrodes on the sides of
the head beeing very close together and electrodes on the top of the
head were very far from each other. Because of that, now the electrodes
are displayed as if the EEG cap was taken off the head and stretched
on a 2D plane.