Tag style names are unique only among the styles of the same type.
This solution was provided in response to trac ticket #167. Some
refactoring was also applied: all styles from a given StylesTagSet are now
contained in an instance of a TagStyles class which also serves all
operations on tag styles. Because tag styles names are not unique
among all tag styles any more, the information containing
the type of the each tag must be also saved to file to allow
to connect tag styles with appropriate tags unambiguously.