Removed the legacy tags export option.
Legacy tags format became incompatible with Svarog
tags representation - legacy channel tags could not
be set for multiple channels (which is done in Svarog
by setting the tag's channel number to -1) or there
can not be tags with equal names but different
types which is now possible in Svarog. This resulted
in errors while exporting Svarog tags and that is why
we decided to remove this functionality.