util/generated-sources: update script to try origin/branchname
When the upstream repository is cloned, origin/generated-sources
tracking branch is created, but not the local generated-sources
branch. This is the usual case, and somebody trying to compile
svarog will likely not have have the local branch, unless they
intend to update it, which is the rare case. Thus, is makes sense
to try origin/generated-sources.

Also, the script is made slightly more robust and now won't fail if
protoc is not available.