pom: make protoc invocation optional
If protoc is not available, it is not invoked. This currently means that
the build will fail with undecipherable error messages, unless the
generated files are already present. They can be "created" with
util/generateds-sources, but this is not done automatically, and no hint
is given. I'm not sure how to do it reasonably.

To compile without protoc:
1) run util/generated-sources
2) compile normally