EEG systems are identified by both symbol and type.
The symbol is - for example - 'EEG 10_20 Cap25' and type
is 'children'/'adults' or some other string which
can be used to further distinguish between one system
or another. Montage generators which do have their
EEG system symbol but do not have the EEG system type
defined are connected with every EEG system of the same
symbol. This allows to create one generator file for
all EEG system identified by the same symbol.