Raise exception if no plugin I18n resources
Previously, null was returned. The exception should not be thrown in the
constructor, because not every plugin _has_ i18n resources. Now loading the
resources is delayed until the first use (the first class to getI18nAccess()),
and if the resources are indeed missing, the exception is not propagated to the
caller (the plugin).

It might be nicer to have the exception thrown directly in the constructor or
during plugin registration, but the plugins would need to specify if i18n
resources are needed.