Repaired the FFT filter functionality.
As a side-effect: changed the FFT library from JTransforms
to apache.commons.math.transform - the latter is capable of returning the
results of the FFT in form of an array of Complex numbers
(meaning Complex number as objects not as an array of interleaved
doubles where even elements are real and odd are imaginary part
of a complex number - this representation of complex numbers
was too complicated...).

This allowed to revert the changes made to the FFTSampleFilterEngine
while changing from the Flanagan library to the JTransform library
which probably were the source of bad FFT filter behaviour.

Also: added unit tests for FourierTransform and WindowFunction.

The JTransforms library is still used by the FFT plugin - this should
be somehow corrected to have only one version of code performing
the FourierTransform...