Fixed exporting signal when all channels are selected.
This change is connected with the channel signal selection that
can span through all channels (its channel number is -1, and
its length is arbitrary - unlike block or page signal selections).
This type of selection was not handled correctly by the export
signal functionality and it also affected the MP5 decomposition.