Changed EEG channel function to be inmutable.
Channel function for channels with EEG channel function cannot
be changed. Also: there is no possibility for channels with
other channel functions to be changed to an EEG channel.
That means that channel function is equal to EEG
only if the apropriate channel is a part of the current
EEG system.