Substituted ErrorsDialog (which used Spring) with Dialogs.
Dialogs is a static class with methods for showing various
dialogs (warnings, errors, yes/no dialogs). ErrorsDialog
was deleted because it relied on Spring which was previously
used for e.g. i18n and validation framework. So, after a change
to org.xnap.commons.i18n ErrorsDialogs did not work correctly.