The logic of is_calibration() was reversed
A window to fill in the calibration parameters was presented when the
codec was capable of getting the calibration parameters itself.

Setting the calibration values from an external source is still broken --
some generated method is missing -- but it doesn't matter much since
both EASYS and EDF contain calibration information.