Use calibration information from codec when reading samples
This allows the signal to be properly scaled.

Apparently the definitions in codecs have an opposite sense and the signal
must be divided, not multiplied by the gain. Nevertheless, it's easier to do
the reversion in code, not in codecs. This code is supposed to be deleted anyway,
so let's take the easiest route.