Control buttons were removed from the EmbeddedFileChooser.
Previously the Control buttons ('OK' and 'Cancel') were
shown in the embedded file chooser and, when pressed,
tried to call the actions of the parent dialog.
This partially worked on Linux but didn't work at all
on MacOS (NullPointerException), and as these buttons
seemed unnecessary and causing bugs, they were removed.