Renamed plugin events and listeners classes to "Plugin*".
For example: previously there was an interface for a plugin
event called "SvarogTagEvent" which was implemented by
the "TagEventImpl". What is more, apart from the plugin system,
there is already a class called "TagEvent" in Svarog, so
it was rather hard to determine what is exactly going on,
which event are we using at the moment. Now the SvarogTagEvent
is called "PluginTagEvent", TagEventImpl -> "PluginTagEventImpl"
and so on for all events and listeners in the plugin system.