Currently added arrow is selected in the VisualReferenceEditor.
Previously, after adding an arrow it was not automatically
selected which resulted in a situation that the weight
text box seemed not to work (it worked but the user had
to click on an arrow to edit its weight). Automatically
selecting arrows after adding improves usability.
This commit is connected with trac ticket #97.