Removed OpenSignalDescriptor and OpenFileSignalDescriptor.
OpenFileSignalDescriptor contained a RawSignalDescriptor
and a SignalmlDescriptor and a FileOpenSignalMethod field
which determined which descriptor should be used.
OpenSignalDescriptor contained an OpenFileSignalDescriptor
and an ExperimentDescriptor and a SignalSource field
which decided which of those was filled and should be used.
Now, OpenDocumentDescriptor contains one AbstractOpenSignalDescriptor
which is an ExperimentDescriptor, a RawSignalDescriptor or
a SignalmlDescriptor depending on what is the type of signal
that should be opened.