Revert "Do not build installer by default"
This reverts commit 881bb3bff838ed945f9842e6641fb11ad12e49c8.

The problem that maven wanted to always download stuff from the
internet seem to be gone now. Since the installer .exe is now created
in the package phase, it will be useful to build it by default
(building the installer will be tested more, and also it will
be built automatically by jenkins without further ado).