izpack: make linux svarog launcher script work
The installation layout is horrible, far from normal linux
conventions. Nevertheless, the script is installed alongside
the jar file, and should look for it in the same directory
that the script ($0) is located, not the current directory.

Also use 'exec' to destory the shell before launching java.

Also drop the redirects - I don't see any reason why we would
want to squelch messages from svarog this way. Especially
error messages.