Added a combo box for manual selection of file type.
File type loading method can be: AUTODETECTED, RAW or using
a SignalML codec. The value selected in this combo box
determines which method  will be used to load the selected
signal file. If the AUTODETECT value is selected, then Svarog
tries to determine the file type automatically based on
the file extension.