favorites: add directory to favorites when the file is really opened
After the dialog was refactored to new form, favorites list stopped being filled
in. Unfortunately now the time when the file is selected and the time when it
is okeyed by the user - the time when "Finish" is clicked - are quite distant.
Nevertheless, we want to actually only add the dir to history when we actually
open the file, i.e. only when "Finish" is clicked.

To achieve this, onDialogCloseWithOK is forwarded through different levels.