While connecting to a new experiment, waits for it to fully start.
1. While connecting to a new experiment, the ConnectToExperimentWorker
ask the experiment for its amplifier parameters. If, as a response,
it receives a list with properly filled parameters, it assumes the
experiments has been correctly started.

2. The MonitorWorker waits for the first sample with infinite
timeout. This gives openBCI some time to fully start the multiplexer
before Svarog can receive samples.

These fixes are due to the longer startup time of the new openBCI
EEG amplifier drivers.