Set svarog.home property to user configured svarog home
Previously it was set to the $CWD, which didn't make much sense, because the
file could be either created in random places, or the logging file would fail to
be created. In this second situation, the following was thrown:
java.io.FileNotFoundException: /usr/local/svarog/log/signalml_log.txt (No such file or directory)
    at java.io.FileOutputStream.open(Native Method)
    at java.io.FileOutputStream.<init>(FileOutputStream.java:209)
    at java.io.FileOutputStream.<init>(FileOutputStream.java:131)
    at org.apache.log4j.FileAppender.setFile(FileAppender.java:289)
    at org.apache.log4j.FileAppender.activateOptions(FileAppender.java:163)
    at org.apache.log4j.config.PropertySetter.activate(PropertySetter.java:256)
    at org.apache.log4j.config.PropertySetter.setProperties(PropertySetter.java:132)
    at org.apache.log4j.config.PropertySetter.setProperties(PropertySetter.java:96)
    at org.apache.log4j.PropertyConfigurator.parseAppender(PropertyConfigurator.java:654)
    at org.apache.log4j.PropertyConfigurator.parseCategory(PropertyConfigurator.java:612)
    at org.apache.log4j.PropertyConfigurator.configureRootCategory(PropertyConfigurator.java:509)
    at org.apache.log4j.PropertyConfigurator.doConfigure(PropertyConfigurator.java:415)
    at org.apache.log4j.PropertyConfigurator.doConfigure(PropertyConfigurator.java:441)
    at org.apache.log4j.PropertyConfigurator.configure(PropertyConfigurator.java:336)
    at org.springframework.util.Log4jConfigurer.initLogging(Log4jConfigurer.java:70)
    at org.signalml.app.SvarogApplication._run(SvarogApplication.java:359)
    at org.signalml.app.SvarogApplication.run(SvarogApplication.java:303)
    at java.lang.Thread.run(Thread.java:636)

The order of some operations during initalization is modified, to first find the
proper value of $signaml.root (the profile directory chosen by the user), then
initalize logging, using $signalml.root in configuration, and then the debugging
statements are executed. This way debug messages can be properly configured
through log4j configuration.

One remaining problem which is not addressed, is two concurrent instances of
Svarog will try to use the same logfile, resulting in the loss of messages. But
this is an independent thing.