BookPlot 'remembers' the properties that were set for it.
If the user changes the properties for a book plot
(e.g. scale type/palette/if the original signal should be visible)
the next book that will be opened has the same properties
as the last activated book.

Also - changed some of the default values for the book plot
properties.