Shows a dialog whenever OutOfMemoryError happens while drawing a book.
While drawing a book, Svarog allocates a lot of multidimensional
arrays. This causes a java.lang.OutOfMemoryError to be thrown
if too many books are opened and the user tries to open yet
another one. A message dialog implemented here handles
this situation more gracefully and ask the user to close other
books and then reopen the problematic book.