Added a right margin for IIR filtering using filt-filt.
If a signal was filtered using filt-filt with a filter having
a very long impulse response, then the signal was distorted on
the right side of the window. The distortion moved with the
window because only the visible part of the signal is filtered
(+ adding some left margin samples).

Adding some samples from the right margin moved the distortion
from the right-side of the window outside of the visible part).