ExportFilteredSignalSampleSource-first approach to lazy filtering.
Filters the signal while saving the result of each filter to
a file and then filtering it with a second filter while reading
the data from that file.

No signal-growing, initial conditions calculation,
filt-filt filtering were not implemented yet.

This is not connected to the GUI yet - was only tested with
a unit test.