Removed the clearNewSamples method from Changeable SampleSource.
Previously the ChangeableMultichannelSampleSource was resposible
for counting the newly added samples and each request for the number
of newly added samples resulted in zeroing the counter.
Now the ChangeableMultichannelSampleSource counts all the samples
that have been added to it and the requester is responsible for
keeping the track of how much the counter has changed from the last
request.