Hidden channels were unhidden after montage changed.
This bug included also disabling the local scales after
changing the montage. Now the channel visibility and its
local scale is remembered according to the montage channel
name (if in previous montage the channel with the name 'Fpz'
was not visible, in the new one it will not be visible either).

Using this approach we assume that the user will not try to
trick us changing the names of the channels.